requirements.txt
aiohttp==3.9.1
aiohttp-cors==0.7.0
aiokafka==0.8.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.0
croniter==2.0.1
datadis @ git+https://github.com/BeeGroup-cimne/beedis@15c57dc549baee3e1b9ea8547793e3d1247eb8b9
dnspython==2.4.2
duckdb==0.7.1
elementpath==4.1.5
falcon==3.1.1
faust-streaming==0.10.20
frozenlist==1.4.0
greenlet==3.0.1
happybase==1.2.0
idna==3.4
intervaltree==3.1.0
isodate==0.6.1
jsonpath-python==1.0.6
kafka-python==2.0.2
mode-streaming==0.3.5
morph_kgc==2.6.4
multidict==6.0.4
mypy-extensions==1.0.0
Naked==0.1.32
neo4j==5.14.1
numpy==1.26.2
opentracing==2.4.0
packaging==23.2
pandas==2.1.3
ply==3.11
pycryptodome==3.19.0
pymongo==4.6.0
pyoxigraph==0.3.20
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.5.2
rdflib==7.0.0
requests==2.31.0
ruamel.yaml==0.18.0
ruamel.yaml.clib==0.2.8
shellescape==3.8.1
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.5
sql_metadata==2.10.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
terminaltables==3.1.10
thefuzz==0.20.0
thriftpy2==0.4.17
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
venusian==3.1.0
yarl==1.9.3
pyhive==0.7.0
thrift==0.16.0
thrift-sasl==0.4.3
